[mapleall]Internal commit msg:     [maplebe]CodeCheck maple_be

+++ src/mapleall/maple_be/include/cg/aarch64/aarch64_insn.h
@@ -318,7 +318,7 @@ class A64OpndEmitVisitor : public OpndEmitVisitor {
 
  private:
   void EmitVectorOperand(const RegOperand &v);
-  void EmitIntReg(const RegOperand &v, uint8 opndSz = kMaxSimm32);
+  void EmitIntReg(const RegOperand &v, int32 opndSz = kMaxSimm32);
 
   const OpndProp *opndProp;
 };
