+++ src/mapleall/maple_be/src/cg/aarch64/aarch64_memlayout.cpp
@@ -497,7 +497,7 @@ void AArch64MemLayout::AssignSpillLocationsToPseudoRegisters() {
   }
   AArch64CGFunc *aarchCGFunc = static_cast<AArch64CGFunc*>(cgFunc);
   RegOperand &baseOpnd = aarchCGFunc->GetOrCreateStackBaseRegOperand();
-  int32 offset = static_cast<int32>(segLocals.GetSize());
+  uint32 offset = segLocals.GetSize();
 
   OfstOperand *offsetOpnd =
       &aarchCGFunc->CreateOfstOpnd(offset + k16BitSize, k64BitSize);
